better ux for arbitrary + auto rlp and compact tests
