+++ crates/primitives/src/transaction/tx_type.rs
@@ -1,8 +1,8 @@
-use reth_codecs::{derive_compact_arbitrary, Compact};
+use reth_codecs::{derive_arbitrary, Compact};
 use serde::{Deserialize, Serialize};
 
 /// Transaction Type
-#[derive_compact_arbitrary]
+#[derive_arbitrary(compact)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Default, Serialize, Deserialize)]
 pub enum TxType {
     /// Legacy transaction pre EIP-2929
